Move test for svp fund tx registration to BridgeSupportRegisterBtcTransactionTest class. Minor refactor to reuse bridgeContractAddress
Convert global variable into local one to fix sonar complain
