+++ rskj-core/src/test/java/co/rsk/peg/BridgeSupportTest.java
@@ -549,11 +549,7 @@ class BridgeSupportTest {
         private final Coin spendableValueFromProposedFederation = bridgeMainNetConstants.getSpendableValueFromProposedFederation();
         private final List<LogInfo> logs = new ArrayList<>();
 
-        private Repository repository;
         private BridgeStorageProvider bridgeStorageProvider;
-
-        private PartialMerkleTree pmtWithTransactions;
-        private int btcBlockWithPmtHeight;
 
         private BridgeSupport bridgeSupport;
 
@@ -601,7 +597,7 @@ class BridgeSupportTest {
                 signatureCache
             );
 
-            repository = createRepository();
+            Repository repository = createRepository();
             bridgeStorageProvider = new BridgeStorageProvider(
                 repository,
                 contractAddress,
@@ -808,259 +804,6 @@ class BridgeSupportTest {
                 .filter(log -> Arrays.equals(log.getData(), expectedData))
                 .findFirst();
             assertTrue(data.isPresent());
-        }
-
-        @Test
-        void registerBtcTransaction_forSvpFundTransactionChange_whenProposedFederationDoesNotExist_shouldRegisterTransactionButNotUpdateSvpFundTransactionValues() throws Exception {
-            // arrange
-            BtcTransaction svpFundTransaction = arrangeSvpFundTransactionWithChange();
-
-            signInputs(svpFundTransaction); // a transaction trying to be registered should be signed
-            setUpForTransactionRegistration(svpFundTransaction);
-
-            when(federationSupport.getProposedFederation()).thenReturn(Optional.empty());
-
-            // act
-            int activeFederationUtxosSizeBeforeRegisteringTx = federationSupport.getActiveFederationBtcUTXOs().size();
-            bridgeSupport.registerBtcTransaction(
-                rskTx,
-                svpFundTransaction.bitcoinSerialize(),
-                btcBlockWithPmtHeight,
-                pmtWithTransactions.bitcoinSerialize()
-            );
-            bridgeStorageProvider.save();
-
-            // assert
-            assertActiveFederationUtxosSize(activeFederationUtxosSizeBeforeRegisteringTx + 1);
-            assertTransactionWasProcessed(svpFundTransaction.getHash());
-            assertSvpFundTransactionValuesWereNotUpdated();
-        }
-
-        @Test
-        void registerBtcTransaction_forSvpFundTransactionChange_whenValidationPeriodEnded_shouldRegisterTransactionButNotUpdateSvpFundTransactionValues() throws Exception {
-            // arrange
-            // make rsk execution block to be after validation period ended
-            long validationPeriodEndBlock = proposedFederation.getCreationBlockNumber()
-                + bridgeMainNetConstants.getFederationConstants().getValidationPeriodDurationInBlocks();
-            long rskExecutionBlockTimestamp = 10L;
-            BlockHeader blockHeader = new BlockHeaderBuilder(mock(ActivationConfig.class))
-                .setNumber(validationPeriodEndBlock + 1) // adding one more block to ensure validation period is ended
-                .setTimestamp(rskExecutionBlockTimestamp)
-                .build();
-            rskExecutionBlock = Block.createBlockFromHeader(blockHeader, true);
-
-            BtcTransaction svpFundTransaction = arrangeSvpFundTransactionWithChange();
-            signInputs(svpFundTransaction); // a transaction trying to be registered should be signed
-            setUpForTransactionRegistration(svpFundTransaction);
-
-            // Act
-            int activeFederationUtxosSizeBeforeRegisteringTx = federationSupport.getActiveFederationBtcUTXOs().size();
-            bridgeSupport.registerBtcTransaction(
-                rskTx,
-                svpFundTransaction.bitcoinSerialize(),
-                btcBlockWithPmtHeight,
-                pmtWithTransactions.bitcoinSerialize()
-            );
-            bridgeStorageProvider.save();
-
-            // assert
-            assertActiveFederationUtxosSize(activeFederationUtxosSizeBeforeRegisteringTx + 1);
-            assertTransactionWasProcessed(svpFundTransaction.getHash());
-            assertSvpFundTransactionValuesWereNotUpdated();
-        }
-
-        @Test
-        void registerBtcTransaction_forNormalPegout_whenSvpPeriodIsOngoing_shouldRegisterTransactionButNotUpdateSvpFundTransactionValues() throws Exception {
-            // Arrange
-            arrangeSvpFundTransactionWithChange();
-
-            BtcTransaction pegout = createPegout();
-            savePegoutIndex(pegout);
-            signInputs(pegout); // a transaction trying to be registered should be signed
-            setUpForTransactionRegistration(pegout);
-
-            // Act
-            int activeFederationUtxosSizeBeforeRegisteringTx = federationSupport.getActiveFederationBtcUTXOs().size();
-            bridgeSupport.registerBtcTransaction(
-                rskTx,
-                pegout.bitcoinSerialize(),
-                btcBlockWithPmtHeight,
-                pmtWithTransactions.bitcoinSerialize()
-            );
-            bridgeStorageProvider.save();
-
-            // assert
-            assertActiveFederationUtxosSize(activeFederationUtxosSizeBeforeRegisteringTx + 1);
-            assertTransactionWasProcessed(pegout.getHash());
-            assertSvpFundTransactionValuesWereNotUpdated();
-        }
-
-        @Test
-        void registerBtcTransaction_forSvpFundTransactionWithoutChangeOutput_whenSvpPeriodIsOngoing_shouldJustUpdateSvpFundTransactionValues() throws Exception {
-            // Arrange
-            BtcTransaction svpFundTransaction = createSvpFundTransaction();
-            saveSvpFundTransactionUnsigned(svpFundTransaction);
-
-            signInputs(svpFundTransaction); // a transaction trying to be registered should be signed
-            setUpForTransactionRegistration(svpFundTransaction);
-
-            // Act
-            int activeFederationUtxosSizeBeforeRegisteringTx = federationSupport.getActiveFederationBtcUTXOs().size();
-            bridgeSupport.registerBtcTransaction(
-                rskTx,
-                svpFundTransaction.bitcoinSerialize(),
-                btcBlockWithPmtHeight,
-                pmtWithTransactions.bitcoinSerialize()
-            );
-            bridgeStorageProvider.save();
-
-            // Assert
-            assertActiveFederationUtxosSize(activeFederationUtxosSizeBeforeRegisteringTx);
-            assertTransactionWasProcessed(svpFundTransaction.getHash());
-            assertSvpFundTransactionValuesWereUpdated();
-        }
-
-        @Test
-        void registerBtcTransaction_forSvpFundTransactionChange_whenSvpPeriodIsOngoing_shouldRegisterTransactionAndUpdateSvpFundTransactionValues() throws Exception {
-            // Arrange
-            BtcTransaction svpFundTransaction = arrangeSvpFundTransactionWithChange();
-
-            signInputs(svpFundTransaction); // a transaction trying to be registered should be signed
-            setUpForTransactionRegistration(svpFundTransaction);
-
-            // Act
-            int activeFederationUtxosSizeBeforeRegisteringTx = federationSupport.getActiveFederationBtcUTXOs().size();
-            bridgeSupport.registerBtcTransaction(
-                rskTx,
-                svpFundTransaction.bitcoinSerialize(),
-                btcBlockWithPmtHeight,
-                pmtWithTransactions.bitcoinSerialize()
-            );
-            bridgeStorageProvider.save();
-
-            // Assert
-            assertActiveFederationUtxosSize(activeFederationUtxosSizeBeforeRegisteringTx + 1);
-            assertTransactionWasProcessed(svpFundTransaction.getHash());
-            assertSvpFundTransactionValuesWereUpdated();
-        }
-
-        private BtcTransaction arrangeSvpFundTransactionWithChange() {
-            BtcTransaction svpFundTransaction = createSvpFundTransaction();
-            addOutputChange(svpFundTransaction);
-            saveSvpFundTransactionUnsigned(svpFundTransaction);
-
-            return svpFundTransaction;
-        }
-
-        private BtcTransaction createSvpFundTransaction() {
-            BtcTransaction svpFundTransaction = new BtcTransaction(btcMainnetParams);
-
-            Sha256Hash parentTxHash = BitcoinTestUtils.createHash(1);
-            addInput(svpFundTransaction, parentTxHash);
-
-            svpFundTransaction.addOutput(spendableValueFromProposedFederation, proposedFederation.getAddress());
-            Address flyoverProposedFederationAddress =
-                PegUtils.getFlyoverAddress(btcMainnetParams, bridgeMainNetConstants.getProposedFederationFlyoverPrefix(), proposedFederation.getRedeemScript());
-            svpFundTransaction.addOutput(spendableValueFromProposedFederation, flyoverProposedFederationAddress);
-
-            return svpFundTransaction;
-        }
-
-        private void saveSvpFundTransactionUnsigned(BtcTransaction svpFundTransaction) {
-            savePegoutIndex(svpFundTransaction);
-            saveSvpFundTransactionHashUnsigned(svpFundTransaction.getHash());
-        }
-
-        private BtcTransaction createPegout() {
-            BtcTransaction pegout = new BtcTransaction(btcMainnetParams);
-            Sha256Hash parentTxHash = BitcoinTestUtils.createHash(2);
-            addInput(pegout, parentTxHash);
-            addOutputChange(pegout);
-
-            return pegout;
-        }
-
-        private void addInput(BtcTransaction transaction, Sha256Hash parentTxHash) {
-            // we need to add an input that we can actually sign,
-            // and we know the private keys for the following scriptSig
-            Federation federation = P2shErpFederationBuilder.builder().build();
-            Script scriptSig = federation.getP2SHScript().createEmptyInputScript(null, federation.getRedeemScript());
-
-            transaction.addInput(parentTxHash, 0, scriptSig);
-        }
-
-        private void signInputs(BtcTransaction transaction) {
-            List<BtcECKey> keysToSign =
-                BitcoinTestUtils.getBtcEcKeysFromSeeds(new String[]{"member01", "member02", "member03", "member04", "member05"}, true);
-            List<TransactionInput> inputs = transaction.getInputs();
-            for (TransactionInput input : inputs) {
-                BitcoinTestUtils.signTransactionInputFromP2shMultiSig(transaction, inputs.indexOf(input), keysToSign);
-            }
-        }
-
-        private void addOutputChange(BtcTransaction transaction) {
-            // add output to the active fed
-            Script activeFederationP2SHScript = activeFederation.getP2SHScript();
-            transaction.addOutput(Coin.COIN.multiply(10), activeFederationP2SHScript);
-        }
-
-        private void savePegoutIndex(BtcTransaction pegout) {
-            BitcoinUtils.getFirstInputSigHash(pegout)
-                .ifPresent(inputSigHash -> bridgeStorageProvider.setPegoutTxSigHash(inputSigHash));
-        }
-
-        private void saveSvpFundTransactionHashUnsigned(Sha256Hash svpFundTransactionHashUnsigned) {
-            bridgeStorageProvider.setSvpFundTxHashUnsigned(svpFundTransactionHashUnsigned);
-            bridgeSupport.save();
-        }
-
-        private void setUpForTransactionRegistration(BtcTransaction transaction) throws BlockStoreException {
-            // recreate a valid chain that has the tx, so it passes the previous checks in registerBtcTransaction
-            BtcBlockStoreWithCache.Factory btcBlockStoreFactory = new RepositoryBtcBlockStoreWithCache.Factory(btcMainnetParams, 100, 100);
-            BtcBlockStoreWithCache btcBlockStoreWithCache = btcBlockStoreFactory.newInstance(repository, bridgeMainNetConstants, bridgeStorageProvider, activationsAfterForks);
-
-            pmtWithTransactions = createValidPmtForTransactions(Collections.singletonList(transaction.getHash()), btcMainnetParams);
-            btcBlockWithPmtHeight = bridgeMainNetConstants.getBtcHeightWhenPegoutTxIndexActivates() + bridgeMainNetConstants.getPegoutTxIndexGracePeriodInBtcBlocks(); // we want pegout tx index to be activated
-
-            int chainHeight = btcBlockWithPmtHeight + bridgeMainNetConstants.getBtc2RskMinimumAcceptableConfirmations();
-            recreateChainFromPmt(btcBlockStoreWithCache, chainHeight, pmtWithTransactions, btcBlockWithPmtHeight, btcMainnetParams);
-
-            bridgeStorageProvider.save();
-
-            bridgeSupport = bridgeSupportBuilder
-                .withBridgeConstants(bridgeMainNetConstants)
-                .withProvider(bridgeStorageProvider)
-                .withActivations(activationsAfterForks)
-                .withFederationSupport(federationSupport)
-                .withFeePerKbSupport(feePerKbSupport)
-                .withExecutionBlock(rskExecutionBlock)
-                .withBtcBlockStoreFactory(btcBlockStoreFactory)
-                .withRepository(repository)
-                .build();
-        }
-
-        private void assertActiveFederationUtxosSize(int expectedActiveFederationUtxosSize) {
-            assertEquals(expectedActiveFederationUtxosSize, federationSupport.getActiveFederationBtcUTXOs().size());
-        }
-
-        private void assertTransactionWasProcessed(Sha256Hash transactionHash) throws IOException {
-            Optional<Long> rskBlockHeightAtWhichBtcTxWasProcessed = bridgeStorageProvider.getHeightIfBtcTxhashIsAlreadyProcessed(transactionHash);
-            assertTrue(rskBlockHeightAtWhichBtcTxWasProcessed.isPresent());
-
-            long rskExecutionBlockNumber = rskExecutionBlock.getNumber();
-            assertEquals(rskExecutionBlockNumber, rskBlockHeightAtWhichBtcTxWasProcessed.get());
-        }
-
-        private void assertSvpFundTransactionValuesWereNotUpdated() {
-            assertTrue(bridgeStorageProvider.getSvpFundTxHashUnsigned().isPresent());
-            assertFalse(bridgeStorageProvider.getSvpFundTxSigned().isPresent());
-        }
-
-        private void assertSvpFundTransactionValuesWereUpdated() {
-            Optional<BtcTransaction> svpFundTransactionSignedOpt = bridgeStorageProvider.getSvpFundTxSigned();
-            assertTrue(svpFundTransactionSignedOpt.isPresent());
-
-            assertFalse(bridgeStorageProvider.getSvpFundTxHashUnsigned().isPresent());
         }
     }
